requirements.txt
psd-tools>=1.10.8
Pillow>=11.2.1
numpy>=1.26.4
